pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Giovanna Manfredo <giovanna.candido.manfredo@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
mysqlclient = "^2.2.4"
python-dotenv = "^1.0.1"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
